R script assignment

Assign a R script or expression to a R symbol in the current R session.

opal.assign.script(opal, symbol, value, async = FALSE)

Arguments

opal
Opal object or list of opal objects.

symbol
Name of the R symbol.

value
The R expression to assign.

async
R script is executed asynchronously within the session (default is FALSE). If TRUE, the value returned is the ID of the command to look for (from Opal 2.1).

See also

Examples

if (FALSE) {
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
# assign a function and call it
opal.assign.script(o, 'hello', quote(function(x) { print(paste0('Hello ', x , '!'))}))
opal.execute(o, "hello('Mr Bean')")
opal.logout(o)
}