Unzip a zip archive file

Unzip a zip archive file from the Opal file system.

opal.file_unzip(opal, source, destination, key = NULL)

Arguments

opal
Opal object.

source
Path to the file in the Opal file system (must exist and have the ".zip" file extension).

destination
Path to the destination file or folder in the Opal file system.

key
Key to decrypt archive.

Value

The path of the extracted archive folder in the Opal file system.

See also

Examples

if (FALSE) {
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
# unzip
path <- opal.file_unzip(o, "/tmp/TESTING.zip", "/home/administrator")
opal.logout(o)
}