Create a resource reference in a project

Create a resource reference in a project

opal.resource_create(
  opal,
  project,
  name,
  url,
  description = NULL,
  format = NULL,
  package = NULL,
  identity = NULL,
  secret = NULL
)

Arguments

opal
Opal object.

project
Name of the project.

name
Name of the resource in the project.

url
The URL of the resource.

description
The description of the resource (optional).

format
The format of the data described by the resource (optional).

package
The R package to be loaded prior to the assignment of the resource (optional).

identity
The identity key or username to be used when accessing the resource (optional).

secret
The secret key or password to be used when accessing the resource (optional).

See also

Examples

if (FALSE) {
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
opal.resource_create(o, 'RSRC', 'CNSIM4', 
  url = 'opal+https://opal-demo.obiba.org/ws/files/projects/RSRC/CNSIM3.zip', 
  format = 'csv', secret = 'EeTtQGIob6haio5bx6FUfVvIGkeZJfGq')
opal.logout(o)
}