Update the dictionary of a Opal table

Directly update the dictionary of a Opal table with the provided dictionary.

Usage

opal.table_dictionary_update(
  opal,
  project,
  table,
  variables,
  categories = NULL,
  complete = FALSE
)

Arguments

opal

Opal connection object.

project

Project name where the table will be located.

table

Destination table name.

variables

A data frame with one row per variable (column name) and then one column per property/attribute (Opal Excel format).

categories

A data frame with one row per category (columns variable and name) and then column per property/attribute (Opal Excel format). If there are no categories, this parameter is optional.

complete

A logical to indicate if the provided dictionary is complete, i.e. that the variables not included in this dictionary must be removed from the view. Otherwise (default) only the addition or the update of variables is performed.

See also

Examples

if (FALSE) { # \dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
variables <- tibble::tribble(
  ~name, ~valueType, ~`label:en`,  ~`Namespace::Name`, ~unit, ~repeatable, ~index,
  "mpg", "decimal", "Mpg label",  "Value1", "years", 0, 1,
  "cyl", "decimal", "Cyl label",  "Value2", "kg/m2", 0, 2,
  "disp", "decimal", "Disp label", NA, NA, 1, 3
)
categories <- tibble::tribble(
  ~variable, ~name, ~missing, ~`label:en`, ~`label:fr`,
  "cyl", "4", 0, "Four", "Quatre",
  "cyl", "6", 0, "Six", "Six",
  "cyl", "8", 1, "Height", "Huit"
)
opal.table_dictionary_update(o, "test", "mtcars", variables, categories)
opal.logout(o)
} # }