Create an Opal view over tables

Create an Opal view if a table with same name does not already exist. The view created will have no dictionary, use opal.table_dictionary_update to apply a dictionary.

Usage

opal.table_view_create(opal, project, table, tables, type = "Participant")

Arguments

opal

Opal connection object.

project

Project name where the table will be located.

table

Table name to be created

tables

List of the fully qualified table names that are referred by the view.

type

Entity type, default is "Participant". Ignored if some table references are provided.

See also

Examples

if (FALSE) { # \dontrun{
o <- opal.login('administrator','password', url='https://opal-demo.obiba.org')
# make a view
opal.table_view_create(o, "CNSIM", "CNSIM123", 
                       c("CNSIM.CNSIM1", "CNSIM.CNSIM2", "CNSIM.CNSIM3"))
opal.logout(o)
} # }